+++ src/router/routes/tyreImpressions.js
@@ -1,0 +1,7 @@
+export default [
+  {
+    path: '/tyre-impression',
+    name: "tyre-impression-list",
+    component: () => import('@/views/tyre_impressions/TyreImpressionList.vue'),
+  },
+]