+++ src/router/routes/tyreModels.js
@@ -1,7 +1,7 @@
 export default [
   {
-    path: '/tyre-models',
-    name: "tyre-models",
+    path: '/tyre-model',
+    name: "tyre-model-list",
     component: () => import('@/views/tyre_models/TyreModelList.vue'),
   },
 ]